Move _compare_version test
