+++ nilearn/_utils/tests/test_helpers.py
@@ -1,3 +1,4 @@
+import pytest
 import warnings
 
 from nilearn._utils import helpers
@@ -126,3 +127,20 @@ def test_future_warn_deprecated_params():
     ):
         assert raised_warning_.category is FutureWarning
         assert str(raised_warning_.message) == expected_warning_
+
+
+@pytest.mark.parametrize("version_a,operator,version_b",
+                         [('0.1.0', '>', '0.0.1'),
+                          ('0.1.0', '>=', '0.0.1'),
+                          ('0.1', '==', '0.1.0'),
+                          ('0.0.0', '<', '0.1.0'),
+                          ('1.0', '!=', '0.1.0')])
+def test_compare_version(version_a, operator, version_b):
+    assert helpers._compare_version(version_a, operator, version_b)
+
+
+def test_compare_version_error():
+    with pytest.raises(ValueError,
+                       match=("'_compare_version' received an "
+                              "unexpected operator <>.")):
+        helpers._compare_version('0.1.0', '<>', '1.1.0')
